requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-tenants==3.7
psycopg2-binary==2.9.10
python-dotenv==1.0.1